move masking specialization to its own header
