+++ include/ck/tensor_operation/gpu/device/gemm_specialization.hpp
@@ -53,23 +53,6 @@ inline std::string getGemmSpecializationString(const GemmSpecialization& s)
     }
 }
 
-// TODO ANT: move to its own file
-enum struct MaskingSpecialization
-{
-    MaskDisabled,
-    MaskOutUpperTriangle
-};
-
-inline std::string getMaskingSpecializationString(const MaskingSpecialization& s)
-{
-    switch(s)
-    {
-    case MaskingSpecialization::MaskDisabled: return "MaskDisabled";
-    case MaskingSpecialization::MaskOutUpperTriangle: return "MaskOutUpperTriangle";
-    default: return "Unrecognized specialization!";
-    }
-}
-
 } // namespace device
 } // namespace tensor_operation
 } // namespace ck
